Move token.ts to lexer/ dir
