+++ src/formatter/ExpressionFormatter.ts
@@ -3,7 +3,7 @@ import { equalizeWhitespace } from 'src/utils';
 
 import Params from 'src/formatter/Params';
 import { isTabularStyle } from 'src/core/config';
-import { isReserved, type Token, TokenType } from 'src/core/token';
+import { isReserved, type Token, TokenType } from 'src/lexer/token';
 import {
   AllColumnsAsterisk,
   ArraySubscript,
