+++ src/formatter/tabularStyle.ts
@@ -1,5 +1,5 @@
 import type { IndentStyle } from 'src/types';
-import { Token, TokenType } from 'src/core/token';
+import { Token, TokenType } from 'src/lexer/token';
 
 /**
  * When tabular style enabled,
