+++ src/lexer/Tokenizer.ts
@@ -1,4 +1,4 @@
-import { Token, TokenType } from 'src/core/token';
+import { Token, TokenType } from 'src/lexer/token';
 import * as regex from 'src/lexer/regexFactory';
 import * as regexTypes from 'src/lexer/regexTypes';
