+++ src/lexer/token.ts
