+++ src/parser/Parser.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-cond-assign */
-import { EOF_TOKEN, type Token, TokenType, isToken } from 'src/core/token';
+import { EOF_TOKEN, type Token, TokenType, isToken } from 'src/lexer/token';
 import {
   AllColumnsAsterisk,
   ArraySubscript,
